requirements.txt
# ---------------- Core ----------------
Flask==3.0.0
Flask-Cors==4.0.0
gunicorn==21.2.0

# ---------------- Auth ----------------
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# ---------------- Database ----------------
sqlalchemy>=2.0.44
psycopg2-binary 


# ---------------- AI / ML ----------------
faiss-cpu==1.13.2
numpy
scikit-learn

# ---------------- Utils ----------------
python-dotenv==1.0.0
requests==2.31.0

groq 
PyPDF2 